Mosaic_Augmentation/requirements.txt
opencv-python>=4.0.0.0
Pillow
numpy